requirements.txt
pytube==15.0.0
openai-whisper==20231117
moviepy==1.0.3
pydub==0.25.1
langchain==0.1.5
chromadb==0.4.22
ollama==0.1.6
sentence-transformers==2.3.0
pyttsx3==2.90
numpy==1.24.4
faiss-cpu==1.7.4
python-dotenv==1.0.0
tqdm==4.66.1